Create testing for Discover Movies Page and add to routes the movies path for user
