+++ config/routes.rb
@@ -7,6 +7,7 @@ Rails.application.routes.draw do
 
   resources :users, only: [:index, :create, :new, :show] do
     resources :user_parties, only: :index, path: 'discover', as: 'discover'
+    resources :movies, only: :index, as: 'results'
   end
 
   get '/register', to: 'users#new', as: 'register'
